added http return status

+++ src/main/java/com/tang/contactservice/controller/ContactServiceController.java
@@ -5,7 +5,9 @@ import java.util.List;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -15,48 +17,77 @@ import org.springframework.web.bind.annotation.RestController;
 import com.tang.contactservice.model.Contact;
 import com.tang.contactservice.service.ContactService;
 
-
 /**
  * Controller to expose methods as web service
  */
 @RestController
-@RequestMapping(value="/contacts/")
+@RequestMapping(value = "/contacts/")
 public class ContactServiceController {
-	
-    public static final Logger LOG = LoggerFactory.getLogger(ContactServiceController.class);
 
-	//used to call CRUD operations
+	public static final Logger LOG = LoggerFactory.getLogger(ContactServiceController.class);
+	public static final String USER_ALREADY_EXISTS = "User Already Exists";
+	public static final String NO_USER_FOUND = "No user found";
+	public static final String USER_DELETED = "User deleted";
+
+	// used to call CRUD operations
 	@Autowired
 	private ContactService contactService;
-	
+
 	// Get all Contacts
-	@RequestMapping(method=RequestMethod.GET, produces=MediaType.APPLICATION_JSON_VALUE)
-	public List<Contact> getAllContacts(){
-		return contactService.getAllContacts();
+	@SuppressWarnings({ "unchecked", "rawtypes" })
+	@RequestMapping(method = RequestMethod.GET, produces = MediaType.APPLICATION_JSON_VALUE)
+	public ResponseEntity<List<Contact>> getAllContacts() {
+		List<Contact> list = contactService.getAllContacts();
+		if(list.isEmpty()){
+			return new ResponseEntity(NO_USER_FOUND, HttpStatus.NOT_FOUND);
+		}
+		return new ResponseEntity<List<Contact>>(list, HttpStatus.OK);
 	}
-	
+
 	// Get contact(s) by name
-	@RequestMapping(value="{name}", method=RequestMethod.GET, produces=MediaType.APPLICATION_JSON_VALUE)
-	public List<Contact> searchContactByName(@PathVariable("name") String name){
-		return contactService.searchContactByName(name);
+	@SuppressWarnings({ "rawtypes", "unchecked" })
+	@RequestMapping(value = "{name}", method = RequestMethod.GET, produces = MediaType.APPLICATION_JSON_VALUE)
+	public ResponseEntity<List<Contact>> searchContactByName(@PathVariable("name") String name) {
+		List<Contact> list = contactService.searchContactByName(name);
+		if(list.isEmpty()){
+			return new ResponseEntity(NO_USER_FOUND, HttpStatus.NOT_FOUND);
+		}
+		return new ResponseEntity<List<Contact>>(list, HttpStatus.OK);
 	}
-	
+
 	// Add a new contact
-	@RequestMapping(method=RequestMethod.POST, consumes=MediaType.APPLICATION_JSON_VALUE)
-	public Contact addContact(@RequestBody Contact contact) throws Exception{
-		LOG.info("adding contact...");
-		return contactService.addContact(contact);
+	@SuppressWarnings("rawtypes")
+	@RequestMapping(method = RequestMethod.POST, consumes = MediaType.APPLICATION_JSON_VALUE)
+	public ResponseEntity addContact(@RequestBody Contact contact){
+		Contact result = contactService.addContact(contact);
+		if (result != null) {
+			return new ResponseEntity<Contact>(result, HttpStatus.CREATED);
+		}
+		return new ResponseEntity<String>(USER_ALREADY_EXISTS, HttpStatus.CONFLICT);
 	}
-	
+
 	// Delete a contact by ID
-	@RequestMapping(value="{id}", method=RequestMethod.DELETE)
-	public void removeContactById(@PathVariable("id") long id) throws Exception{
-		contactService.removeContactById(id);
+	@SuppressWarnings("rawtypes")
+	@RequestMapping(value = "{id}", method = RequestMethod.DELETE)
+	public ResponseEntity removeContactById(@PathVariable("id") long id) {
+		if(contactService.findContactById(id)){
+			contactService.removeContactById(id);
+			return new ResponseEntity<String>(USER_DELETED, HttpStatus.OK);
+		}
+		return new ResponseEntity<String>(NO_USER_FOUND, HttpStatus.NOT_FOUND);
 	}
-	
+
 	// Modify a contact by ID
-	@RequestMapping(value="{id}", method=RequestMethod.PUT, consumes=MediaType.APPLICATION_JSON_VALUE)
-	public Contact modifyContactById(@PathVariable("id") long id, @RequestBody Contact contact) throws Exception{
-		return contactService.modifyContactById(id, contact);
+	@SuppressWarnings("rawtypes")
+	@RequestMapping(value = "{id}", method = RequestMethod.PUT, consumes = MediaType.APPLICATION_JSON_VALUE)
+	public ResponseEntity modifyContactById(@PathVariable("id") long id, @RequestBody Contact contact) {
+		if(contactService.findContactById(id)){
+			Contact result = contactService.modifyContactById(id, contact);
+			if (result != null) {
+				return new ResponseEntity<Contact>(result, HttpStatus.CREATED);
+			}
+			return new ResponseEntity<String>(USER_ALREADY_EXISTS, HttpStatus.CONFLICT);
+		}
+		return new ResponseEntity<String>(NO_USER_FOUND, HttpStatus.NOT_FOUND);
 	}
 }
